ieee_ras_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-socketio
fastapi-socketio
python-multipart
aiosqlite
python-socketio[asyncio_client]
PyJWT>=2.0.0
